reduce grap producerl
