+++ jnosql-mapping/jnosql-mapping-graph/src/test/java/org/eclipse/jnosql/mapping/graph/GraphTemplateProducerTest.java
@@ -25,7 +25,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 @CDIExtension
-public class DefaultGraphTemplateProducerTest {
+public class GraphTemplateProducerTest {
 
     @Inject
     private GraphTemplateProducer producer;
